Goodbye, C code! 🎉
