+++ Sources/NIOHTTP2/HTTP2Error.swift
@@ -12,8 +12,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-import CNIONghttp2
-
 public protocol NIOHTTP2Error: Equatable, Error { }
 
 /// Errors that NIO raises when handling HTTP/2 connections.
@@ -50,14 +48,6 @@ public enum NIOHTTP2Errors {
 
     public struct BadClientMagic: NIOHTTP2Error {
         public init() {}
-    }
-
-    public struct InternalError: NIOHTTP2Error {
-        internal var nghttp2ErrorCode: nghttp2_error
-
-        internal init(nghttp2ErrorCode: nghttp2_error) {
-            self.nghttp2ErrorCode = nghttp2ErrorCode
-        }
     }
 
     /// A stream state transition was attempted that was not valid.
